+++ HISTORY.md
@@ -50,6 +50,7 @@
 ### 2019-03-07
 * AudioComp lib: added block to set the microphone's buffer and fft sizes
 * German translation update (microphone features)
+* simplified "globe" symbol
 
 ### 2019-03-06
 * AudioComp lib: turn off mic after 5 secs of idling